Address my own feedback
