+++ rest/proto/version.txt
@@ -1,1 +1,1 @@
-v0.75.0-rc.3
+v0.75.0-rc.4
